WIP trying to make my 6261-based GT08 watch into a flashrom serprog device...
The SPI seems to work (i get a valid looking RDID) but the serial port just
seemst to hang after the fernly usb loader is done ... also, it doesnt
do the shell i asked for, need to look into that next.

frser.c is based on libfrser. I'm the author of that code and
license this subset under the LICENSE file of fernly.